backend/requirements.txt
# Har Har Mahadev
fastapi
uvicorn
jpype1
numpy
scipy
scikit-fmm
celery
redis
psycopg2-binary
python-multipart
pyyaml